Core: Merge tlb implementations
